requirements.txt
absl-py==1.4.0
ansi2html==1.8.0
anyio==3.6.1
apache-beam==2.46.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==23.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Brotli==1.0.9
cachetools==5.3.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
contextlib2==21.6.0
crcmod==1.7
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.0.0
fastavro==1.8.3
fasteners==0.18
fastjsonschema==2.16.1
Flask==2.2.2
Flask-Compress==1.12
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-vizier==0.1.11
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-tools==1.48.2
h5py==3.9.0
hdfs==2.7.2
httplib2==0.21.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-dash==0.4.2
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
keras==2.13.1
libclang==16.0.6
lxml==4.9.1
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
ml-collections==0.1.1
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.3.0
orjson==3.9.7
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
portpicker==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pymongo==3.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.20
stack-data==0.5.0
tenacity==8.0.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-serving-api==2.13.0
termcolor==2.3.0
terminado==0.15.0
tinycss2==1.1.1
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.66.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.15.0
zipp==3.8.1
zstandard==0.21.0
